feat: Add fetchHotels function to API client
The fetchHotels function is added to the API client to fetch all hotels from the API. This function makes a GET request to the `/api/hotels` endpoint and returns a Promise that resolves to an array of HotelType objects. If the response is not successful, an error is thrown.

This commit addresses the need to fetch all hotels from the API and provides a convenient function for doing so.